Import Illumina manifests (sesameData versions)

Load either the HM540 and EPIC manifests into working memory

ImportSesameData(manifest_char)

Arguments

manifest_char

Which manifest should be loaded? Currently, this package has been tested to work with 450k and EPIC arrays for HG19 and HG38.

Details

This function assumes that the .onLoad() function has executed properly and (therefore) that the necessary data sets are in the cache.

Examples

  hm450k_gr <- ImportSesameData("HM450.hg19.manifest")
#> snapshotDate(): 2021-05-18
#> see ?sesameData and browseVignettes('sesameData') for documentation
  head(hm450k_gr)
#> GRanges object with 6 ranges and 52 metadata columns:
#>              seqnames      ranges strand | address_A address_B     channel
#>                 <Rle>   <IRanges>  <Rle> | <integer> <integer> <character>
#>   cg13869341     chr1 15865-15866      - |  62703328  16661461         Red
#>   cg14008030     chr1 18827-18828      - |  27651330      <NA>        Both
#>   cg12045430     chr1 29407-29408      - |  25703424  34666387         Red
#>   cg20826792     chr1 29425-29426      - |  61731400  14693326         Red
#>   cg00381604     chr1 29435-29436      - |  26752380  50693408         Red
#>   cg20253340     chr1 68849-68850      - |  49752305      <NA>        Both
#>               designType    nextBase nextBaseRef   probeType orientation
#>              <character> <character> <character> <character> <character>
#>   cg13869341           I           A           C          cg        down
#>   cg14008030          II         G/A           C          cg        down
#>   cg12045430           I           A           C          cg        down
#>   cg20826792           I           T           A          cg        down
#>   cg00381604           I           A           T          cg        down
#>   cg20253340          II         G/A           C          cg        down
#>              probeCpGcnt context35  probeBeg  probeEnd             ProbeSeq_A
#>                <integer> <integer> <integer> <numeric>            <character>
#>   cg13869341           2         4     15865     15914 CCAATAACTAACCACTCTAC..
#>   cg14008030           2         3     18828     18877 ACTCRAAATTTACTCAATAA..
#>   cg12045430           8        11     29407     29456 AAAAAAAACACAATAAAAAA..
#>   cg20826792           7        12     29425     29474 AATCCTAAAACCACACTCAA..
#>   cg00381604           7        12     29435     29484 AAATCAACAAAATCCTAAAA..
#>   cg20253340           1         1     68850     68899 ATTCTATCTCRCACTACTTC..
#>                          ProbeSeq_B                   gene
#>                         <character>            <character>
#>   cg13869341 CCGATAACTAACCACTCTAC..                 WASH7P
#>   cg14008030                              MIR6859-3;WASH7P
#>   cg12045430 GAAAAAAACGCAATAAAAAA.. MIR1302-9;RP11-34P13..
#>   cg20826792 AATCCTAAAACCGCGCTCGA.. MIR1302-9;RP11-34P13..
#>   cg00381604 GAATCGACGAAATCCTAAAA.. MIR1302-9;RP11-34P13..
#>   cg20253340                                         OR4F5
#>                           gene_HGNC      chrm_A     beg_A    flag_A    mapQ_A
#>                         <character> <character> <integer> <integer> <integer>
#>   cg13869341                 WASH7P        chr1     15865        16        17
#>   cg14008030       MIR6859-3;WASH7P        chr1     18828        16         1
#>   cg12045430 MIR1302-9;RP11-34P13..        chr1     29407        16         4
#>   cg20826792 MIR1302-9;RP11-34P13..        chr1     29425        16         4
#>   cg00381604 MIR1302-9;RP11-34P13..        chr1     29435        16         4
#>   cg20253340                  OR4F5        chr1     68850        16         6
#>                  cigar_A      NM_A      chrm_B     beg_B    flag_B    mapQ_B
#>              <character> <integer> <character> <integer> <integer> <integer>
#>   cg13869341         50M         0        chr1     15865        16        17
#>   cg14008030         50M         0        <NA>      <NA>      <NA>      <NA>
#>   cg12045430         50M         0        chr1     29407        16         4
#>   cg20826792         50M         0        chr1     29425        16         4
#>   cg00381604         50M         0        chr1     29435        16         4
#>   cg20253340         50M         0        <NA>      <NA>      <NA>      <NA>
#>                  cigar_B      NM_B wDecoy_chrm_A wDecoy_beg_A wDecoy_flag_A
#>              <character> <integer>   <character>    <integer>     <integer>
#>   cg13869341         50M         0          chr1        15865            16
#>   cg14008030        <NA>      <NA>          chr1        18828            16
#>   cg12045430         50M         0          chr1        29407            16
#>   cg20826792         50M         0          chr1        29425            16
#>   cg00381604         50M         0          chr1        29435            16
#>   cg20253340        <NA>      <NA>          chr1        68850            16
#>              wDecoy_mapQ_A wDecoy_cigar_A wDecoy_NM_A wDecoy_chrm_B
#>                  <integer>    <character>   <integer>   <character>
#>   cg13869341            17            50M           0          chr1
#>   cg14008030             1            50M           0          <NA>
#>   cg12045430             4            50M           0          chr1
#>   cg20826792             4            50M           0          chr1
#>   cg00381604             4            50M           0          chr1
#>   cg20253340             6            50M           0          <NA>
#>              wDecoy_beg_B wDecoy_flag_B wDecoy_mapQ_B wDecoy_cigar_B
#>                 <integer>     <integer>     <integer>    <character>
#>   cg13869341        15865            16            17            50M
#>   cg14008030         <NA>          <NA>          <NA>           <NA>
#>   cg12045430        29407            16             4            50M
#>   cg20826792        29425            16             4            50M
#>   cg00381604        29435            16             4            50M
#>   cg20253340         <NA>          <NA>          <NA>           <NA>
#>              wDecoy_NM_B  posMatch MASK_mapping MASK_typeINextBaseSwitch
#>                <integer> <logical>    <logical>                <logical>
#>   cg13869341           0      TRUE         TRUE                    FALSE
#>   cg14008030        <NA>      TRUE         TRUE                    FALSE
#>   cg12045430           0      TRUE         TRUE                    FALSE
#>   cg20826792           0      TRUE         TRUE                    FALSE
#>   cg00381604           0      TRUE         TRUE                    FALSE
#>   cg20253340        <NA>      TRUE         TRUE                    FALSE
#>              MASK_rmsk15 MASK_sub40_copy MASK_sub35_copy MASK_sub30_copy
#>                <logical>       <logical>       <logical>       <logical>
#>   cg13869341       FALSE            TRUE            TRUE            TRUE
#>   cg14008030       FALSE           FALSE           FALSE           FALSE
#>   cg12045430       FALSE           FALSE            TRUE            TRUE
#>   cg20826792       FALSE           FALSE           FALSE           FALSE
#>   cg00381604       FALSE           FALSE           FALSE           FALSE
#>   cg20253340       FALSE            TRUE            TRUE            TRUE
#>              MASK_sub25_copy MASK_snp5_common MASK_snp5_GMAF1p MASK_extBase
#>                    <logical>        <logical>        <logical>    <logical>
#>   cg13869341            TRUE            FALSE            FALSE        FALSE
#>   cg14008030           FALSE            FALSE            FALSE        FALSE
#>   cg12045430            TRUE            FALSE            FALSE        FALSE
#>   cg20826792           FALSE            FALSE            FALSE        FALSE
#>   cg00381604            TRUE            FALSE            FALSE        FALSE
#>   cg20253340            TRUE            FALSE            FALSE        FALSE
#>              MASK_general
#>                 <logical>
#>   cg13869341         TRUE
#>   cg14008030         TRUE
#>   cg12045430         TRUE
#>   cg20826792         TRUE
#>   cg00381604         TRUE
#>   cg20253340         TRUE
#>   -------
#>   seqinfo: 26 sequences from an unspecified genome; no seqlengths