Split CpG dataframe by Subregion

Split a dataframe of CpGs and comethylated subregions to a list of CpGs in each subregion

SplitCpGDFbyRegion(
  CpGsSubregions_df,
  genome = c("hg19", "hg38"),
  arrayType = c("450k", "EPIC"),
  manifest_gr = NULL,
  returnAllCpGs = TRUE
)

Arguments

CpGsSubregions_df

data frame with CpG and subregion number

genome

Human genome of reference: hg19 or hg38

arrayType

Type of array: 450k or EPIC

manifest_gr

A GRanges object with the genome manifest (as returned by ExperimentHub or by ImportSesameData). This function by default ignores this argument in favour of the genome and arrayType arguments.

returnAllCpGs

indicates if outputting all the CpGs in the region when there is not a contiguous comethylated region or only the CpGs in the contiguous comethylated regions

Value

a list of comethylated subregions CpGs for a pre-defined region

Examples

   data(betaMatrix_ex4)
   CpGs_df <- MarkComethylatedCpGs(betaCluster_mat = betaMatrix_ex4)
   CpGsSubregions_df <- FindComethylatedRegions(CpGs_df)

   SplitCpGDFbyRegion(
     CpGsSubregions_df,
     genome = "hg19",
     arrayType = "450k"
   )
#> snapshotDate(): 2021-05-18
#> see ?sesameData and browseVignettes('sesameData') for documentation
#> snapshotDate(): 2021-05-18
#> see ?sesameData and browseVignettes('sesameData') for documentation
#> snapshotDate(): 2021-05-18
#> see ?sesameData and browseVignettes('sesameData') for documentation
#> $`chr7:27182637-27183794`
#>  [1] "cg05076221" "cg23936031" "cg02248486" "cg25866143" "cg09549073"
#>  [6] "cg01370449" "cg04863892" "cg19759481" "cg12128839" "cg02916332"
#> [11] "cg17569124" "cg02005600" "cg25307665" "cg14014955" "cg02646423"
#> 
#> $`chr7:27183816-27184264`
#>  [1] "cg23204968" "cg05835726" "cg14058329" "cg03207666" "cg13512268"
#>  [6] "cg23454797" "cg12015737" "cg08070327" "cg25506432" "cg14658493"
#> [11] "cg19196335" "cg09880291" "cg24389585" "cg16997642" "cg20817131"
#> [16] "cg14013695" "cg25390165" "cg01323381"
#> 
#> $`chr7:27184316-27185512`
#>  [1] "cg05774699" "cg26023912" "cg14882265" "cg17432857" "cg00969405"
#>  [6] "cg07049592" "cg02106682" "cg03368099" "cg01748892" "cg13694927"
#> [11] "cg03744763" "cg27151303" "cg05579037" "cg21641458" "cg09343092"
#> [16] "cg18931036" "cg24040595"
#>